*update enkf.nml.tmp file to remove "use_qsatensmean" that was abandoned in GSI code;  and leave pseudo_rh as an option for enkf_mean/enkf_recenter or enfk_update;  add corresponding condition in exhafs_enkf.sh; *turn on ps and use tsen instead of tv for enkf_mean/enkf_recenter in anavinfo.tmp_enkf; *modify the exhafs_forecast.sh to avoid mv existing files error in RESTART file. *update submodule hafs_gsi.fd

+++ scripts/exhafs_forecast.sh
@@ -1426,7 +1426,9 @@ dynhrs=$(echo $CDATEnhrs | cut -c7-8)
 hhnhrs=$(echo ${CDATEnhrs} | cut -c9-10)
 if [ -s ${YMDnhrs}.${hhnhrs}*.fv_core.res.nc ]; then
   for file in $(/bin/ls -1 ${YMDnhrs}.${hhnhrs}*.fv*.nc* ${YMDnhrs}.${hhnhrs}*.phy_data*.nc* ${YMDnhrs}.${hhnhrs}*.sfc_data*.nc* ${YMDnhrs}.${hhnhrs}*.coupler.res); do
-    mv ${file} ${YMDnhrs}.${hhnhrs}0000.${file:16}
+    if [ ! -s ${YMDnhrs}.${hhnhrs}0000.${file:16} ]; then
+      mv ${file} ${YMDnhrs}.${hhnhrs}0000.${file:16}
+    fi
   done
   if [ ${RUN_INIT:-NO} = YES ]; then
     sed -i -e "3s/.*/  ${yrnhrs}    $(echo ${mnnhrs}|sed 's/^0/ /')    $(echo ${dynhrs}|sed 's/^0/ /')    $(echo ${hhnhrs}|sed 's/^0/ /')     0     0        Current model time: year, month, day, hour, minute, second/" ${YMDnhrs}.${hhnhrs}0000.coupler.res
